+++ src/navigation/stacks/RecordStackNavigation.tsx
@@ -4,13 +4,13 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 import { RecordNavigationProps } from '../types/navigationTypes';
 
-import AchievementEditView from '@/screens/Record/AchievementEditView';
-import CareerEditView from '@/screens/Record/CareerEditView';
-import EducationEditView from '@/screens/Record/EducationEditView';
-import LanguageEditView from '@/screens/Record/LanguageEditView';
+import AchievementEditView from '@/screens/Record/EditView/AchievementEditView';
+import CareerEditView from '@/screens/Record/EditView/CareerEditView';
+import EducationEditView from '@/screens/Record/EditView/EducationEditView';
+import LanguageEditView from '@/screens/Record/EditView/LanguageEditView';
+import ProjectEditView from '@/screens/Record/EditView/ProjectEditView';
+import ResumeEditView from '@/screens/Record/EditView/ResumeEditView';
 import LinkEditView from '@/screens/Record/LinkEditView';
-import ProjectEditView from '@/screens/Record/ProjectEditView';
-import ResumeEditView from '@/screens/Record/ResumeEditView';
 
 const Stack = createNativeStackNavigator<RecordNavigationProps>();
