+++ src/screens/Record/EditView/EducationEditView.tsx
